% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-js-computed.R
\name{alpine_compute}
\alias{alpine_compute}
\title{Computed Property: Custom Expression}
\usage{
alpine_compute(expr)
}
\arguments{
\item{expr}{Character string: JavaScript expression that can reference
\code{this} and other state. Will be wrapped in a function that returns the result.}
}
\value{
An \code{htmlwidgets::JS} object containing a JavaScript function.
}
\description{
Helper to create a computed property from an arbitrary JavaScript expression.
Use for computed properties that don't fit the standard filter/sum patterns.
}
\details{
This is the escape hatch for computed properties that need custom logic.
The expression should be valid JavaScript that can reference \code{this}:
\itemize{
\item \code{"this.totalIncome() - this.totalExpense()"} -- arithmetic
\item \code{"this.users.length > 0"} -- boolean check
\item \code{"this.items.filter(x => x.active).map(x => x.name)"} -- transformation
}
}
\examples{
\donttest{
htmltools::tags$div() |>
  alpine_data(
    totalIncome = 5000,
    totalExpense = 2000,
    balance = alpine_compute("this.totalIncome - this.totalExpense"),
    hasItems = alpine_compute("this.balance > 0"),
    netPercent = alpine_compute("((this.totalIncome - this.totalExpense) / this.totalIncome * 100).toFixed(2)")
  )
}

}
\seealso{
Other Computed Properties: 
\code{\link{alpine_combine_conditions}()},
\code{\link{alpine_count_filtered}()},
\code{\link{alpine_multi_condition}()},
\code{\link{alpine_sum_filtered}()}
}
\concept{Computed Properties}
