% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magics.R
\name{alpine_el}
\alias{alpine_el}
\title{Access the current DOM element}
\usage{
alpine_el()
}
\value{
An \code{htmlwidgets::JS} object containing \code{"$el"}.
}
\description{
Returns the \verb{$el} magic property string, which refers to the root DOM element
of the current Alpine.js component. Use it in expressions wherever a direct
reference to the component's root node is needed.
}
\details{
\verb{$el} is available in all Alpine expressions and refers to the element that
has \code{x-data}. Common uses include reading dimensions, scrolling the component
into view, or passing the element to a third-party library.
}
\examples{
\donttest{
# Scroll the component into view on a button click
htmltools::tags$button("Scroll Here") |>
  alpine_on("click", paste0(alpine_el(), ".scrollIntoView()"))
}

}
\seealso{
Other Magic Properties: 
\code{\link{alpine_dispatch}()},
\code{\link{alpine_next_tick}()},
\code{\link{alpine_refs}()},
\code{\link{alpine_store_js}()},
\code{\link{alpine_watch}()}
}
\concept{Magic Properties}
