% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{alpine_ignore}
\alias{alpine_ignore}
\title{Prevent Alpine.js from initializing a block of HTML}
\usage{
alpine_ignore(tag, .cssSelector = NULL)
}
\arguments{
\item{tag}{An htmltools tag object.}

\item{.cssSelector}{Optional CSS selector for targeting a nested element.}
}
\value{
The modified tag with \code{x-ignore} attribute appended.
}
\description{
Adds the \code{x-ignore} attribute, which tells Alpine.js to skip the element
and all of its children during initialization. Useful for embedding
third-party widgets or code that should not be processed by Alpine.
}
\examples{
\donttest{
htmltools::tags$div(
  htmltools::tags$p("This paragraph uses x-text.") |> alpine_text("msg"),
  htmltools::tags$div("Third-party widget -- Alpine will leave this alone.") |>
    alpine_ignore()
) |>
  alpine_data(msg = "Hello")
}

}
\seealso{
Other Core Directives: 
\code{\link{alpine_bind}()},
\code{\link{alpine_cloak}()},
\code{\link{alpine_data}()},
\code{\link{alpine_data_js}()},
\code{\link{alpine_effect}()},
\code{\link{alpine_for}()},
\code{\link{alpine_html}()},
\code{\link{alpine_id}()},
\code{\link{alpine_if}()},
\code{\link{alpine_init}()},
\code{\link{alpine_modelable}()},
\code{\link{alpine_on}()},
\code{\link{alpine_ref}()},
\code{\link{alpine_show}()},
\code{\link{alpine_teleport}()},
\code{\link{alpine_text}()},
\code{\link{alpine_transition}()},
\code{\link{alpine_when}()}
}
\concept{Core Directives}
