% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{alpine_init}
\alias{alpine_init}
\title{Execute a JavaScript expression on component initialization}
\usage{
alpine_init(tag, expression, .cssSelector = NULL)
}
\arguments{
\item{tag}{An htmltools tag object.}

\item{expression}{Character string containing JavaScript code to execute.}

\item{.cssSelector}{Optional CSS selector for targeting a nested element.}
}
\value{
The modified tag with \code{x-init} attribute appended.
}
\description{
Creates an \code{x-init} attribute containing JavaScript code that Alpine.js executes
after the component's reactive state is initialized. Useful for loading data,
setting up event listeners, or running one-time setup code.
}
\details{
Expression runs after \code{x-data} initialization.
}
\examples{
\donttest{
htmltools::tags$div() |>
  alpine_init("loadData()") |>
  alpine_data(data = list())
}

}
\seealso{
Other Core Directives: 
\code{\link{alpine_bind}()},
\code{\link{alpine_cloak}()},
\code{\link{alpine_data}()},
\code{\link{alpine_data_js}()},
\code{\link{alpine_effect}()},
\code{\link{alpine_for}()},
\code{\link{alpine_html}()},
\code{\link{alpine_id}()},
\code{\link{alpine_if}()},
\code{\link{alpine_ignore}()},
\code{\link{alpine_modelable}()},
\code{\link{alpine_on}()},
\code{\link{alpine_ref}()},
\code{\link{alpine_show}()},
\code{\link{alpine_teleport}()},
\code{\link{alpine_text}()},
\code{\link{alpine_transition}()},
\code{\link{alpine_when}()}
}
\concept{Core Directives}
