% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{alpine_is_plugin_registered}
\alias{alpine_is_plugin_registered}
\title{Check if a Plugin is Registered}
\usage{
alpine_is_plugin_registered(plugin_name)
}
\arguments{
\item{plugin_name}{Character string naming the plugin.}
}
\value{
Logical; \code{TRUE} if registered, \code{FALSE} otherwise.
}
\description{
Quickly check whether a named plugin has been registered.
}
\examples{
\donttest{
alpine_register_plugin("persist")
alpine_is_plugin_registered("persist")
}

}
\seealso{
Other Plugin System: 
\code{\link{alpine_get_plugin_registry}()},
\code{\link{alpine_register_plugin}()},
\code{\link{plugin_templates}}
}
\concept{Plugin System}
