% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-js-storage.R
\name{alpine_storage_set}
\alias{alpine_storage_set}
\title{Set Value in localStorage}
\usage{
alpine_storage_set(key, value_expr = "this")
}
\arguments{
\item{key}{Character string: the localStorage key to set}

\item{value_expr}{Character string: JavaScript expression to store.
Default is \code{"this"} (store the entire component state).
Can reference any state property: \code{"this.todos"}, \code{"this.user.name"}, etc.}
}
\value{
A character string containing a JavaScript statement for use in
event handlers. Must be passed to functions that accept JS statements
like \code{js_event_handler()} or event handlers like \verb{\\@click}.
}
\description{
Helper to create a JavaScript statement that stores a value in localStorage.
Typically used within event handlers for persistence.
}
\details{
This helper generates:

\if{html}{\out{<div class="sourceCode javascript">}}\preformatted{localStorage.setItem('key', JSON.stringify(\{value_expr\}));
}\if{html}{\out{</div>}}

For use in event handlers, pass this result to an \verb{\\@event} attribute or
include in a composite event handler.
}
\examples{
\donttest{
htmltools::tags$div(
  htmltools::tags$input(type = "text", placeholder = "Todo") |> alpine_model("newTodo"),
  htmltools::tags$button("Add & Save") |>
    alpine_on("click", 
      "todos.push({id: Date.now(), text: newTodo}); localStorage.setItem('todos', JSON.stringify(todos)); newTodo = ''"),
  htmltools::tags$ul(
    alpine_for(
      htmltools::tags$li() |> alpine_text("item.text"),
      "item in todos",
      key = "item.id"
    )
  )
) |>
  alpine_data(
    todos = alpine_storage_get("todos", default = list()),
    newTodo = ""
  )
}

}
\seealso{
Other Client Storage: 
\code{\link{alpine_storage_get}()}
}
\concept{Client Storage}
