% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad hook for alpinejs}
\usage{
.onLoad(libname, pkgname)
}
\description{
Sets default package options and initializes the plugin registry when alpinejs is loaded.
}
\keyword{internal}
